Refactor email verification module into DDD layers
